Fix Metadata contract bug
* In AvastarMetadata.sol,
  - in getAvastarMetadata,
     - change uint8 to uint256 for token id

* Updated AvastarMetadata.flat.sol

* Updated migrate_partial/2_deploy_contracts.js
  - only deploy and prepare the new AvastarMetadata contract

* In start-ganache.js,
  - changed gas limit

* In truffle.js,
  - changed gas price in all environments to 12 gwei

+++ contracts/AvastarMetadata.sol
@@ -213,7 +213,7 @@ contract AvastarMetadata is AvastarBase, AvastarTypes, AccessControl {
         attribution = strConcat('Original art by: ', attribution);
 
         // Name
-        metadata = strConcat('{\n  "name": "Avastar #', uintToStr(uint8(id)));
+        metadata = strConcat('{\n  "name": "Avastar #', uintToStr(uint256(id)));
         metadata = strConcat(metadata, '",\n');
 
         // Description: Generation
